Refactor feature testing utilities
