+++ src/testFixtures/java/org/opensearch/knn/FeatureTestCase.java
@@ -1,10 +1,986 @@
+/*
+ * Copyright OpenSearch Contributors
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
 package org.opensearch.knn;
 
-/**
- * Generic test case that exposes the utilities in {@link DerivedSourceTestCase} for
- * features beyond derived source. Extend this class when adding integration tests
- * for new features that require comprehensive index operations.
- */
-public abstract class FeatureTestCase extends DerivedSourceTestCase {
-    // Intentionally empty. Inherit functionality from DerivedSourceTestCase.
+import lombok.SneakyThrows;
+import org.apache.hc.core5.http.io.entity.EntityUtils;
+import org.opensearch.client.Request;
+import org.opensearch.client.Response;
+import org.opensearch.common.xcontent.XContentFactory;
+import org.opensearch.core.rest.RestStatus;
+import org.opensearch.core.xcontent.XContentBuilder;
+import org.opensearch.index.query.MatchAllQueryBuilder;
+import org.opensearch.index.query.QueryBuilder;
+import org.opensearch.knn.index.VectorDataType;
+import org.opensearch.knn.IDVectorProducer;
+import org.opensearch.knn.TestUtils;
+
+import java.io.IOException;
+import java.util.ArrayList;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Locale;
+import java.util.Map;
+import java.util.Random;
+import java.util.Set;
+import java.util.function.Supplier;
+
+public class FeatureTestCase extends KNNRestTestCase {
+
+    private static final List<Pair<String, Boolean>> INDEX_PREFIX_TO_ENABLED = List.of(
+        new Pair<>("original-enable-", true),
+        new Pair<>("original-disable-", false),
+        new Pair<>("e2e-", true),
+        new Pair<>("e2d-", false),
+        new Pair<>("d2e-", true),
+        new Pair<>("d2d-", false)
+    );
+
+    private static final int MIN_DIMENSION = 4;
+    private static final int MAX_DIMENSION = 32;
+    private static final int MIN_DOCS = 50;
+    private static final int MAX_DOCS = 200;
+
+    /**
+     * Testing flat, single field base case with index configuration. The test will automatically skip adding fields for
+     *  random documents to ensure it works robustly. To ensure correctness, we repeat same operations against an
+     *  index without derived source enabled (baseline).
+     *  {
+     *     "settings": {
+     *         "index.knn" true,
+     *         "index.knn.derived_source.enabled": true/false
+     *     },
+     *     "mappings":{
+     *         "properties": {
+     *             "test_float_vector": {
+     *                 "type": "knn_vector",
+     *                 "dimension": 16,
+     *                 "data_type": float
+     *             },
+     *  			"test_float_vector": {
+     *                 "type": "knn_vector",
+     *                 "dimension": 16,
+     *                 "data_type": float
+     *             },
+     * 			"test_float_vector": {
+     *                 "type": "knn_vector",
+     *                 "dimension": 16,
+     *                 "data_type": float
+     *             },
+     *             "test_float_vector": {
+     *                 "type": "knn_vector",
+     *                 "dimension": 16,
+     *                 "data_type": float
+     *             },
+     *             "test_float_vector": {
+     *                 "type": "knn_vector",
+     *                 "dimension": 16,
+     *                 "data_type": float
+     *             },
+     *             "test_float_vector": {
+     *                 "type": "knn_vector",
+     *                 "dimension": 16,
+     *                 "data_type": float
+     *             },
+     *             "test-text" {
+     *             	"type": "text"
+     *             },
+     *             "test-int" {
+     *             	"type": "text"
+     *             },
+     *         }
+     *     }
+     * }
+     */
+    protected List<FeatureTestUtils.IndexConfigContext> getFlatIndexContexts(String testSuitePrefix, boolean addRandom, boolean addNull) {
+        List<FeatureTestUtils.IndexConfigContext> indexConfigContexts = new ArrayList<>();
+        long consistentRandomSeed = random().nextLong();
+        for (Pair<String, Boolean> index : INDEX_PREFIX_TO_ENABLED) {
+            Supplier<Integer> dimensionSupplier = randomIntegerSupplier(consistentRandomSeed, MIN_DIMENSION, MAX_DIMENSION);
+            Supplier<Integer> binaryDimensionSupplier = randomIntegerSupplier(consistentRandomSeed, MIN_DIMENSION, MAX_DIMENSION, 8);
+            Supplier<Integer> randomDocCountSupplier = randomIntegerSupplier(consistentRandomSeed, MIN_DOCS, MAX_DOCS);
+            FeatureTestUtils.IndexConfigContext indexConfigContext = FeatureTestUtils.IndexConfigContext.builder()
+                .indexName(getIndexName(testSuitePrefix, index.getFirst(), addRandom))
+                .docCount(randomDocCountSupplier.get())
+                .derivedEnabled(index.getSecond())
+                .random(new Random(consistentRandomSeed))
+                .fields(
+                    List.of(
+                        FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                            .dimension(dimensionSupplier.get())
+                            .nullProb(addNull ? FeatureTestUtils.DEFAULT_NULL_PROB : 0)
+                            .fieldPath("test_float_vector")
+                            .build(),
+                        FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                            .dimension(dimensionSupplier.get())
+                            .nullProb(addNull ? FeatureTestUtils.DEFAULT_NULL_PROB : 0)
+                            .fieldPath("update_float_vector")
+                            .isUpdate(true)
+                            .build(),
+                        FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                            .fieldPath("test_byte_vector")
+                            .vectorDataType(VectorDataType.BYTE)
+                            .nullProb(addNull ? FeatureTestUtils.DEFAULT_NULL_PROB : 0)
+                            .build(),
+                        FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                            .fieldPath("update_byte_vector")
+                            .vectorDataType(VectorDataType.BYTE)
+                            .dimension(dimensionSupplier.get())
+                            .nullProb(addNull ? FeatureTestUtils.DEFAULT_NULL_PROB : 0)
+                            .isUpdate(true)
+                            .build(),
+                        FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                            .fieldPath("test_binary_vector")
+                            .vectorDataType(VectorDataType.BINARY)
+                            .dimension(binaryDimensionSupplier.get())
+                            .nullProb(addNull ? FeatureTestUtils.DEFAULT_NULL_PROB : 0)
+                            .build(),
+                        FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                            .fieldPath("update_binary_vector")
+                            .vectorDataType(VectorDataType.BINARY)
+                            .dimension(binaryDimensionSupplier.get())
+                            .nullProb(addNull ? FeatureTestUtils.DEFAULT_NULL_PROB : 0)
+                            .isUpdate(true)
+                            .build(),
+                        FeatureTestUtils.TextFieldType.builder().fieldPath("test-text").build(),
+                        FeatureTestUtils.IntFieldType.builder().fieldPath("test-int").build()
+                    )
+                )
+                .build();
+            indexConfigContext.init();
+            indexConfigContexts.add(indexConfigContext);
+        }
+        return indexConfigContexts;
+    }
+
+    /**
+     * Object field
+     * {
+     *     "settings": {
+     *         "index.knn" true,
+     *         "index.knn.derived_source.enabled": true/false
+     *     },
+     *     "mappings" : {
+     *       "properties" : {
+     *         "path_1" : {
+     *           "properties" : {
+     *             "test_vector" : {
+     *               "type" : "knn_vector",
+     *               "dimension" : 63
+     *             },
+     *             "update_vector" : {
+     *               "type" : "knn_vector",
+     *               "dimension" : 34
+     *             }
+     *           }
+     *         },
+     *         "path_2" : {
+     *           "properties" : {
+     *             "path_3" : {
+     *               "properties" : {
+     *                 "test-int" : {
+     *                   "type" : "integer"
+     *                 },
+     *                 "test_vector" : {
+     *                   "type" : "knn_vector",
+     *                   "dimension" : 41
+     *                 },
+     *                 "update_vector" : {
+     *                   "type" : "knn_vector",
+     *                   "dimension" : 8
+     *                 }
+     *               }
+     *             },
+     *             "test-text" : {
+     *               "type" : "text"
+     *             },
+     *             "test_vector" : {
+     *               "type" : "knn_vector",
+     *               "dimension" : 45
+     *             },
+     *             "update_vector" : {
+     *               "type" : "knn_vector",
+     *               "dimension" : 7
+     *             }
+     *           }
+     *         },
+     *         "test-int" : {
+     *           "type" : "integer"
+     *         },
+     *         "test-text" : {
+     *           "type" : "text"
+     *         },
+     *         "test_vector" : {
+     *           "type" : "knn_vector",
+     *           "dimension" : 10
+     *         },
+     *         "update_vector" : {
+     *           "type" : "knn_vector",
+     *           "dimension" : 51
+     *         }
+     *       }
+     *     }
+     *   }
+     * }
+     */
+    protected List<FeatureTestUtils.IndexConfigContext> getObjectIndexContexts(String testSuitePrefix, boolean addRandom) {
+        List<FeatureTestUtils.IndexConfigContext> indexConfigContexts = new ArrayList<>();
+        long consistentRandomSeed = random().nextLong();
+        for (Pair<String, Boolean> index : INDEX_PREFIX_TO_ENABLED) {
+            Supplier<Integer> dimensionSupplier = randomIntegerSupplier(consistentRandomSeed, MIN_DIMENSION, MAX_DIMENSION);
+            Supplier<Integer> randomDocCountSupplier = randomIntegerSupplier(consistentRandomSeed, MIN_DOCS, MAX_DOCS);
+            FeatureTestUtils.IndexConfigContext indexConfigContext = FeatureTestUtils.IndexConfigContext.builder()
+                .indexName(getIndexName(testSuitePrefix, index.getFirst(), addRandom))
+                .docCount(randomDocCountSupplier.get())
+                .derivedEnabled(index.getSecond())
+                .random(new Random(consistentRandomSeed))
+                .fields(
+                    List.of(
+                        FeatureTestUtils.ObjectFieldContext.builder()
+                            .fieldPath("path_1")
+                            .children(
+                                List.of(
+                                    FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                                        .dimension(dimensionSupplier.get())
+                                        .fieldPath("path_1.test_vector")
+                                        .build(),
+                                    FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                                        .dimension(dimensionSupplier.get())
+                                        .fieldPath("path_1.update_vector")
+                                        .isUpdate(true)
+                                        .build()
+                                )
+                            )
+                            .build(),
+                        FeatureTestUtils.ObjectFieldContext.builder()
+                            .fieldPath("path_2")
+                            .children(
+                                List.of(
+                                    FeatureTestUtils.TextFieldType.builder().fieldPath("path_2.test-text").build(),
+                                    FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                                        .dimension(dimensionSupplier.get())
+                                        .fieldPath("path_2.test_vector")
+                                        .build(),
+                                    FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                                        .dimension(dimensionSupplier.get())
+                                        .fieldPath("path_2.update_vector")
+                                        .isUpdate(true)
+                                        .build(),
+                                    FeatureTestUtils.ObjectFieldContext.builder()
+                                        .fieldPath("path_2.path_3")
+                                        .children(
+                                            List.of(
+                                                FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                                                    .dimension(dimensionSupplier.get())
+                                                    .fieldPath("path_2.path_3.test_vector")
+                                                    .build(),
+                                                FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                                                    .dimension(dimensionSupplier.get())
+                                                    .fieldPath("path_2.path_3.update_vector")
+                                                    .isUpdate(true)
+                                                    .build(),
+                                                FeatureTestUtils.IntFieldType.builder().fieldPath("path_2.path_3.test-int").build()
+                                            )
+                                        )
+                                        .build()
+                                )
+                            )
+                            .build(),
+                        FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                            .dimension(dimensionSupplier.get())
+                            .fieldPath("test_vector")
+                            .build(),
+                        FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                            .dimension(dimensionSupplier.get())
+                            .fieldPath("update_vector")
+                            .isUpdate(true)
+                            .build(),
+
+                        FeatureTestUtils.TextFieldType.builder().fieldPath("test-text").build(),
+                        FeatureTestUtils.IntFieldType.builder().fieldPath("test-int").build()
+                    )
+                )
+                .build();
+            indexConfigContext.init();
+            indexConfigContexts.add(indexConfigContext);
+        }
+
+        return indexConfigContexts;
+    }
+
+    /**
+     * Testing nested fields
+     * {
+     *     "settings": {
+     *         "index.knn" true,
+     *         "index.knn.derived_source.enabled": true/false
+     *     },
+     *     "mappings" : {
+     *       "properties" : {
+     *         "nested_1" : {
+     *           "type" : "nested",
+     *           "properties" : {
+     *             "object_1" : {
+     *               "properties" : {
+     *                 "test-int" : {
+     *                   "type" : "integer"
+     *                 },
+     *                 "test_vector" : {
+     *                   "type" : "knn_vector",
+     *                   "dimension" : 64
+     *                 }
+     *               }
+     *             },
+     *             "test_vector" : {
+     *               "type" : "knn_vector",
+     *               "dimension" : 9
+     *             },
+     *             "update_vector" : {
+     *               "type" : "knn_vector",
+     *               "dimension" : 4
+     *             }
+     *           }
+     *         },
+     *         "nested_2" : {
+     *           "type" : "nested",
+     *           "properties" : {
+     *             "nested_3" : {
+     *               "type" : "nested",
+     *               "properties" : {
+     *                 "test-int" : {
+     *                   "type" : "integer"
+     *                 },
+     *                 "test_vector" : {
+     *                   "type" : "knn_vector",
+     *                   "dimension" : 27
+     *                 },
+     *                 "update_vector" : {
+     *                   "type" : "knn_vector",
+     *                   "dimension" : 14
+     *                 }
+     *               }
+     *             },
+     *             "test-text" : {
+     *               "type" : "text"
+     *             },
+     *             "test_vector" : {
+     *               "type" : "knn_vector",
+     *               "dimension" : 57
+     *             },
+     *             "update_vector" : {
+     *               "type" : "knn_vector",
+     *               "dimension" : 10
+     *             }
+     *           }
+     *         },
+     *         "object_1" : {
+     *           "properties" : {
+     *             "nested_1" : {
+     *               "type" : "nested",
+     *               "properties" : {
+     *                 "test_vector" : {
+     *                   "type" : "knn_vector",
+     *                   "dimension" : 30
+     *                 }
+     *               }
+     *             },
+     *             "test_vector" : {
+     *               "type" : "knn_vector",
+     *               "dimension" : 51
+     *             }
+     *           }
+     *         },
+     *         "test-int" : {
+     *           "type" : "integer"
+     *         },
+     *         "test-text" : {
+     *           "type" : "text"
+     *         },
+     *         "test_vector" : {
+     *           "type" : "knn_vector",
+     *           "dimension" : 63
+     *         },
+     *         "update_vector" : {
+     *           "type" : "knn_vector",
+     *           "dimension" : 4
+     *         }
+     *       }
+     *     }
+     *   }
+     */
+    protected List<FeatureTestUtils.IndexConfigContext> getNestedIndexContexts(String testSuitePrefix, boolean addRandom) {
+        List<FeatureTestUtils.IndexConfigContext> indexConfigContexts = new ArrayList<>();
+        long consistentRandomSeed = random().nextLong();
+        for (Pair<String, Boolean> index : INDEX_PREFIX_TO_ENABLED) {
+            Supplier<Integer> dimensionSupplier = randomIntegerSupplier(consistentRandomSeed, MIN_DIMENSION, MAX_DIMENSION);
+            Supplier<Integer> randomDocCountSupplier = randomIntegerSupplier(consistentRandomSeed, MIN_DOCS, MAX_DOCS);
+            FeatureTestUtils.IndexConfigContext indexConfigContext = FeatureTestUtils.IndexConfigContext.builder()
+                .indexName(getIndexName(testSuitePrefix, index.getFirst(), addRandom))
+                .docCount(randomDocCountSupplier.get())
+                .derivedEnabled(index.getSecond())
+                .random(new Random(consistentRandomSeed))
+                .fields(
+                    List.of(
+                        FeatureTestUtils.ObjectFieldContext.builder()
+                            .fieldPath("object_1")
+                            .children(
+                                List.of(
+                                    FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                                        .dimension(dimensionSupplier.get())
+                                        .fieldPath("object_1.test_vector")
+                                        .build(),
+                                    FeatureTestUtils.NestedFieldContext.builder()
+                                        .fieldPath("object_1.nested_1")
+                                        .children(
+                                            List.of(
+                                                FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                                                    .dimension(dimensionSupplier.get())
+                                                    .fieldPath("object_1.nested_1.test_vector")
+                                                    .build()
+                                            )
+                                        )
+                                        .build()
+                                )
+                            )
+                            .build(),
+                        FeatureTestUtils.NestedFieldContext.builder()
+                            .fieldPath("nested_1")
+                            .children(
+                                List.of(
+                                    FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                                        .dimension(dimensionSupplier.get())
+                                        .fieldPath("nested_1.test_vector")
+                                        .build(),
+                                    FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                                        .dimension(dimensionSupplier.get())
+                                        .fieldPath("nested_1.update_vector")
+                                        .isUpdate(true)
+                                        .build(),
+                                    FeatureTestUtils.ObjectFieldContext.builder()
+                                        .fieldPath("nested_1.object_1")
+                                        .children(
+                                            List.of(
+                                                FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                                                    .dimension(dimensionSupplier.get())
+                                                    .fieldPath("nested_1.object_1.test_vector")
+                                                    .build(),
+                                                FeatureTestUtils.IntFieldType.builder().fieldPath("nested_1.object_1.test-int").build()
+                                            )
+                                        )
+                                        .build()
+                                )
+                            )
+                            .build(),
+                        FeatureTestUtils.NestedFieldContext.builder()
+                            .fieldPath("nested_2")
+                            .children(
+                                List.of(
+                                    FeatureTestUtils.TextFieldType.builder().fieldPath("nested_2.test-text").build(),
+                                    FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                                        .dimension(dimensionSupplier.get())
+                                        .fieldPath("nested_2.test_vector")
+                                        .build(),
+                                    FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                                        .fieldPath("nested_2.update_vector")
+                                        .isUpdate(true)
+                                        .dimension(dimensionSupplier.get())
+                                        .build(),
+                                    FeatureTestUtils.NestedFieldContext.builder()
+                                        .fieldPath("nested_2.nested_3")
+                                        .children(
+                                            List.of(
+                                                FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                                                    .dimension(dimensionSupplier.get())
+                                                    .fieldPath("nested_2.nested_3.test_vector")
+                                                    .build(),
+                                                FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                                                    .dimension(dimensionSupplier.get())
+                                                    .fieldPath("nested_2.nested_3.update_vector")
+                                                    .isUpdate(true)
+                                                    .build(),
+                                                FeatureTestUtils.IntFieldType.builder().fieldPath("nested_2.nested_3.test-int").build()
+                                            )
+                                        )
+                                        .build()
+                                )
+                            )
+                            .build(),
+                        FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                            .dimension(dimensionSupplier.get())
+                            .fieldPath("test_vector")
+                            .build(),
+                        FeatureTestUtils.KNNVectorFieldTypeContext.builder()
+                            .dimension(dimensionSupplier.get())
+                            .fieldPath("update_vector")
+                            .isUpdate(true)
+                            .build(),
+                        FeatureTestUtils.TextFieldType.builder().fieldPath("test-text").build(),
+                        FeatureTestUtils.IntFieldType.builder().fieldPath("test-int").build()
+                    )
+                )
+                .build();
+            indexConfigContext.init();
+            indexConfigContexts.add(indexConfigContext);
+        }
+        return indexConfigContexts;
+    }
+
+    @SneakyThrows
+    protected void prepareOriginalIndices(List<FeatureTestUtils.IndexConfigContext> indexConfigContexts) {
+        assertTrue(1 < indexConfigContexts.size());
+        FeatureTestUtils.IndexConfigContext derivedSourceEnabledContext = indexConfigContexts.get(0);
+        FeatureTestUtils.IndexConfigContext derivedSourceDisabledContext = indexConfigContexts.get(1);
+        createKnnIndex(
+            derivedSourceEnabledContext.indexName,
+            derivedSourceEnabledContext.getSettings(),
+            derivedSourceEnabledContext.getMapping()
+        );
+        createKnnIndex(
+            derivedSourceDisabledContext.indexName,
+            derivedSourceDisabledContext.getSettings(),
+            derivedSourceDisabledContext.getMapping()
+        );
+
+        for (int i = 0; i < derivedSourceDisabledContext.docCount; i++) {
+            String doc1 = derivedSourceEnabledContext.buildDoc();
+            String doc2 = derivedSourceDisabledContext.buildDoc();
+            assertEquals(doc1, doc2);
+            addKnnDoc(derivedSourceEnabledContext.getIndexName(), String.valueOf(i + 1), doc1);
+            addKnnDoc(derivedSourceDisabledContext.getIndexName(), String.valueOf(i + 1), doc2);
+        }
+        refreshAllIndices();
+        assertDocsMatch(
+            derivedSourceDisabledContext.docCount,
+            derivedSourceDisabledContext.indexName,
+            derivedSourceEnabledContext.indexName
+        );
+    }
+
+    @SneakyThrows
+    protected void testMerging(List<FeatureTestUtils.IndexConfigContext> indexConfigContexts) {
+        FeatureTestUtils.IndexConfigContext derivedSourceEnabledContext = indexConfigContexts.get(0);
+        FeatureTestUtils.IndexConfigContext derivedSourceDisabledContext = indexConfigContexts.get(1);
+        String originalIndexNameDerivedSourceEnabled = derivedSourceEnabledContext.indexName;
+        String originalIndexNameDerivedSourceDisabled = derivedSourceDisabledContext.indexName;
+        forceMergeKnnIndex(originalIndexNameDerivedSourceEnabled, 10);
+        forceMergeKnnIndex(originalIndexNameDerivedSourceDisabled, 10);
+        refreshAllIndices();
+        assertIndexBigger(originalIndexNameDerivedSourceDisabled, originalIndexNameDerivedSourceEnabled);
+        assertDocsMatch(
+            derivedSourceDisabledContext.docCount,
+            originalIndexNameDerivedSourceDisabled,
+            originalIndexNameDerivedSourceEnabled
+        );
+        refreshAllIndices();
+        forceMergeKnnIndex(originalIndexNameDerivedSourceEnabled, 1);
+        forceMergeKnnIndex(originalIndexNameDerivedSourceDisabled, 1);
+        refreshAllIndices();
+        assertIndexBigger(originalIndexNameDerivedSourceDisabled, originalIndexNameDerivedSourceEnabled);
+        assertDocsMatch(
+            derivedSourceDisabledContext.docCount,
+            originalIndexNameDerivedSourceDisabled,
+            originalIndexNameDerivedSourceEnabled
+        );
+    }
+
+    @SneakyThrows
+    protected void testUpdate(List<FeatureTestUtils.IndexConfigContext> indexConfigContexts) {
+        FeatureTestUtils.IndexConfigContext derivedSourceEnabledContext = indexConfigContexts.get(0);
+        FeatureTestUtils.IndexConfigContext derivedSourceDisabledContext = indexConfigContexts.get(1);
+        String originalIndexNameDerivedSourceEnabled = derivedSourceEnabledContext.indexName;
+        String originalIndexNameDerivedSourceDisabled = derivedSourceDisabledContext.indexName;
+
+        // Update via POST /<index>/_doc/<docid>
+        // For this, we are just going to replace all of the docs.
+        for (int i = 0; i < derivedSourceEnabledContext.docCount; i += 11) {
+            addKnnDoc(derivedSourceEnabledContext.getIndexName(), String.valueOf(i + 1), derivedSourceEnabledContext.buildDoc());
+            addKnnDoc(derivedSourceDisabledContext.getIndexName(), String.valueOf(i + 1), derivedSourceDisabledContext.buildDoc());
+        }
+        refreshAllIndices();
+        assertDocsMatch(
+            derivedSourceDisabledContext.docCount,
+            originalIndexNameDerivedSourceDisabled,
+            originalIndexNameDerivedSourceEnabled
+        );
+
+        // Use update API
+        for (int i = 0; i < derivedSourceEnabledContext.docCount; i += 13) {
+            updateUpdateAPI(
+                derivedSourceEnabledContext.getIndexName(),
+                String.valueOf(i + 1),
+                derivedSourceEnabledContext.partialUpdateSupplier()
+            );
+            updateUpdateAPI(
+                derivedSourceDisabledContext.getIndexName(),
+                String.valueOf(i + 1),
+                derivedSourceDisabledContext.partialUpdateSupplier()
+            );
+        }
+
+        // Update by query
+        for (int i = 0; i < derivedSourceEnabledContext.docCount; i += 17) {
+            updateKnnDocByQuery(
+                derivedSourceEnabledContext.getIndexName(),
+                derivedSourceEnabledContext.updateByQuerySupplier(String.valueOf(i + 1))
+            );
+            updateKnnDocByQuery(
+                derivedSourceDisabledContext.getIndexName(),
+                derivedSourceDisabledContext.updateByQuerySupplier(String.valueOf(i + 1))
+            );
+        }
+
+        // Sets the doc to an empty doc - make sure this is last - it can mess up update by query
+        setDocToEmpty(originalIndexNameDerivedSourceEnabled, String.valueOf(1));
+        setDocToEmpty(originalIndexNameDerivedSourceDisabled, String.valueOf(1));
+        refreshAllIndices();
+        assertDocsMatch(
+            derivedSourceDisabledContext.docCount,
+            originalIndexNameDerivedSourceDisabled,
+            originalIndexNameDerivedSourceEnabled
+        );
+    }
+
+    @SneakyThrows
+    protected void testSearch(List<FeatureTestUtils.IndexConfigContext> indexConfigContexts) {
+        FeatureTestUtils.IndexConfigContext derivedSourceEnabledContext = indexConfigContexts.get(0);
+        String originalIndexNameDerivedSourceEnabled = derivedSourceEnabledContext.indexName;
+
+        // Default - all fields should be there
+        validateSearch(originalIndexNameDerivedSourceEnabled, derivedSourceEnabledContext.docCount, true, null, null);
+
+        // Default - no fields should be there
+        validateSearch(originalIndexNameDerivedSourceEnabled, derivedSourceEnabledContext.docCount, false, null, null);
+
+        // Exclude all vectors
+        validateSearch(
+            originalIndexNameDerivedSourceEnabled,
+            derivedSourceEnabledContext.docCount,
+            true,
+            null,
+            derivedSourceEnabledContext.collectFieldNames()
+        );
+
+        // Include all vectors
+        validateSearch(
+            originalIndexNameDerivedSourceEnabled,
+            derivedSourceEnabledContext.docCount,
+            true,
+            derivedSourceEnabledContext.collectFieldNames(),
+            null
+        );
+    }
+
+    @SneakyThrows
+    protected void validateSearch(String indexName, int size, boolean isSourceEnabled, List<String> includes, List<String> excludes) {
+        // TODO: We need to figure out a way to enhance validation
+        QueryBuilder qb = new MatchAllQueryBuilder();
+        Request request = new Request("POST", "/" + indexName + "/_search");
+
+        request.addParameter("size", Integer.toString(size));
+        XContentBuilder builder = XContentFactory.jsonBuilder().startObject();
+        builder.field("query", qb);
+        if (isSourceEnabled == false) {
+            builder.field("_source", false);
+        }
+        if (includes != null) {
+            builder.startObject("_source");
+            builder.startArray("includes");
+            for (String include : includes) {
+                builder.value(include);
+            }
+            builder.endArray();
+            builder.endObject();
+        }
+        if (excludes != null) {
+            builder.startObject("_source");
+            builder.startArray("excludes");
+            for (String exclude : excludes) {
+                builder.value(exclude);
+            }
+            builder.endArray();
+            builder.endObject();
+        }
+
+        builder.endObject();
+        request.setJsonEntity(builder.toString());
+
+        Response response = client().performRequest(request);
+        assertEquals(request.getEndpoint() + ": failed", RestStatus.OK, RestStatus.fromCode(response.getStatusLine().getStatusCode()));
+
+        String responseBody = EntityUtils.toString(response.getEntity());
+        List<Object> hits = parseSearchResponseHits(responseBody);
+
+        assertNotEquals(0, hits.size());
+    }
+
+    @SneakyThrows
+    protected void testDelete(List<FeatureTestUtils.IndexConfigContext> indexConfigContexts) {
+        int docToDelete = 8;
+        int docToDeleteByQuery = 11;
+
+        FeatureTestUtils.IndexConfigContext derivedSourceEnabledContext = indexConfigContexts.get(0);
+        FeatureTestUtils.IndexConfigContext derivedSourceDisabledContext = indexConfigContexts.get(1);
+        String originalIndexNameDerivedSourceEnabled = derivedSourceEnabledContext.indexName;
+        String originalIndexNameDerivedSourceDisabled = derivedSourceDisabledContext.indexName;
+
+        // Delete by API
+        deleteKnnDoc(originalIndexNameDerivedSourceEnabled, String.valueOf(docToDelete));
+        deleteKnnDoc(originalIndexNameDerivedSourceDisabled, String.valueOf(docToDelete));
+        refreshAllIndices();
+        assertDocsMatch(
+            derivedSourceDisabledContext.docCount,
+            originalIndexNameDerivedSourceDisabled,
+            originalIndexNameDerivedSourceEnabled
+        );
+
+        // Delete by query
+        deleteKnnDocByQuery(originalIndexNameDerivedSourceEnabled, String.valueOf(docToDeleteByQuery));
+        deleteKnnDocByQuery(originalIndexNameDerivedSourceDisabled, String.valueOf(docToDeleteByQuery));
+        refreshAllIndices();
+        assertDocsMatch(
+            derivedSourceDisabledContext.docCount,
+            originalIndexNameDerivedSourceDisabled,
+            originalIndexNameDerivedSourceEnabled
+        );
+    }
+
+    @SneakyThrows
+    protected void testReindex(List<FeatureTestUtils.IndexConfigContext> indexConfigContexts) {
+        FeatureTestUtils.IndexConfigContext derivedSourceEnabledContext = indexConfigContexts.get(0);
+        FeatureTestUtils.IndexConfigContext derivedSourceDisabledContext = indexConfigContexts.get(1);
+        FeatureTestUtils.IndexConfigContext reindexFromEnabledToEnabledContext = indexConfigContexts.get(2);
+        FeatureTestUtils.IndexConfigContext reindexFromEnabledToDisabledContext = indexConfigContexts.get(3);
+        FeatureTestUtils.IndexConfigContext reindexFromDisabledToEnabledContext = indexConfigContexts.get(4);
+        FeatureTestUtils.IndexConfigContext reindexFromDisabledToDisabledContext = indexConfigContexts.get(5);
+
+        String originalIndexNameDerivedSourceEnabled = derivedSourceEnabledContext.indexName;
+        String originalIndexNameDerivedSourceDisabled = derivedSourceDisabledContext.indexName;
+        String reindexFromEnabledToEnabledIndexName = reindexFromEnabledToEnabledContext.indexName;
+        String reindexFromEnabledToDisabledIndexName = reindexFromEnabledToDisabledContext.indexName;
+        String reindexFromDisabledToEnabledIndexName = reindexFromDisabledToEnabledContext.indexName;
+        String reindexFromDisabledToDisabledIndexName = reindexFromDisabledToDisabledContext.indexName;
+
+        createKnnIndex(
+            reindexFromEnabledToEnabledIndexName,
+            reindexFromEnabledToEnabledContext.getSettings(),
+            reindexFromEnabledToEnabledContext.getMapping()
+        );
+        createKnnIndex(
+            reindexFromEnabledToDisabledIndexName,
+            reindexFromEnabledToDisabledContext.getSettings(),
+            reindexFromEnabledToDisabledContext.getMapping()
+        );
+        createKnnIndex(
+            reindexFromDisabledToEnabledIndexName,
+            reindexFromDisabledToEnabledContext.getSettings(),
+            reindexFromDisabledToEnabledContext.getMapping()
+        );
+        createKnnIndex(
+            reindexFromDisabledToDisabledIndexName,
+            reindexFromDisabledToDisabledContext.getSettings(),
+            reindexFromDisabledToDisabledContext.getMapping()
+        );
+        refreshAllIndices();
+        reindex(originalIndexNameDerivedSourceEnabled, reindexFromEnabledToEnabledIndexName);
+        reindex(originalIndexNameDerivedSourceEnabled, reindexFromEnabledToDisabledIndexName);
+        reindex(originalIndexNameDerivedSourceDisabled, reindexFromDisabledToEnabledIndexName);
+        reindex(originalIndexNameDerivedSourceDisabled, reindexFromDisabledToDisabledIndexName);
+
+        // Need to forcemerge before comparison
+        refreshAllIndices();
+        forceMergeKnnIndex(originalIndexNameDerivedSourceEnabled, 1);
+        forceMergeKnnIndex(originalIndexNameDerivedSourceDisabled, 1);
+        refreshAllIndices();
+        assertIndexBigger(originalIndexNameDerivedSourceDisabled, originalIndexNameDerivedSourceEnabled);
+
+        assertIndexBigger(originalIndexNameDerivedSourceDisabled, reindexFromEnabledToEnabledIndexName);
+        assertIndexBigger(originalIndexNameDerivedSourceDisabled, reindexFromDisabledToEnabledIndexName);
+        assertIndexBigger(reindexFromEnabledToDisabledIndexName, originalIndexNameDerivedSourceEnabled);
+        assertIndexBigger(reindexFromDisabledToDisabledIndexName, originalIndexNameDerivedSourceEnabled);
+        assertDocsMatch(
+            derivedSourceDisabledContext.docCount,
+            originalIndexNameDerivedSourceDisabled,
+            reindexFromEnabledToEnabledIndexName
+        );
+        assertDocsMatch(
+            derivedSourceDisabledContext.docCount,
+            originalIndexNameDerivedSourceDisabled,
+            reindexFromDisabledToEnabledIndexName
+        );
+        assertDocsMatch(
+            derivedSourceDisabledContext.docCount,
+            originalIndexNameDerivedSourceDisabled,
+            reindexFromEnabledToDisabledIndexName
+        );
+        assertDocsMatch(
+            derivedSourceDisabledContext.docCount,
+            originalIndexNameDerivedSourceDisabled,
+            reindexFromDisabledToDisabledIndexName
+        );
+    }
+
+    @SneakyThrows
+    protected void testSnapshotRestore(
+        String repository,
+        String snapshot,
+        List<FeatureTestUtils.IndexConfigContext> indexConfigContexts
+    ) {
+        FeatureTestUtils.IndexConfigContext derivedSourceEnabledContext = indexConfigContexts.get(0);
+        FeatureTestUtils.IndexConfigContext derivedSourceDisabledContext = indexConfigContexts.get(1);
+        FeatureTestUtils.IndexConfigContext reindexFromEnabledToEnabledContext = indexConfigContexts.get(2);
+        FeatureTestUtils.IndexConfigContext reindexFromEnabledToDisabledContext = indexConfigContexts.get(3);
+        FeatureTestUtils.IndexConfigContext reindexFromDisabledToEnabledContext = indexConfigContexts.get(4);
+        FeatureTestUtils.IndexConfigContext reindexFromDisabledToDisabledContext = indexConfigContexts.get(5);
+
+        String originalIndexNameDerivedSourceEnabled = derivedSourceEnabledContext.indexName;
+        String originalIndexNameDerivedSourceDisabled = derivedSourceDisabledContext.indexName;
+        String reindexFromEnabledToEnabledIndexName = reindexFromEnabledToEnabledContext.indexName;
+        String reindexFromEnabledToDisabledIndexName = reindexFromEnabledToDisabledContext.indexName;
+        String reindexFromDisabledToEnabledIndexName = reindexFromDisabledToEnabledContext.indexName;
+        String reindexFromDisabledToDisabledIndexName = reindexFromDisabledToDisabledContext.indexName;
+
+        createSnapshot(repository, snapshot, true);
+
+        deleteIndex(originalIndexNameDerivedSourceEnabled);
+        deleteIndex(originalIndexNameDerivedSourceDisabled);
+        deleteIndex(reindexFromEnabledToEnabledIndexName);
+        deleteIndex(reindexFromEnabledToDisabledIndexName);
+        deleteIndex(reindexFromDisabledToEnabledIndexName);
+        deleteIndex(reindexFromDisabledToDisabledIndexName);
+
+        String restoreSuffix = "-restored";
+        restoreSnapshot(
+            restoreSuffix,
+            List.of(
+                originalIndexNameDerivedSourceEnabled,
+                originalIndexNameDerivedSourceDisabled,
+                reindexFromEnabledToEnabledIndexName,
+                reindexFromEnabledToDisabledIndexName,
+                reindexFromDisabledToEnabledIndexName,
+                reindexFromDisabledToDisabledIndexName
+            ),
+            repository,
+            snapshot,
+            true
+        );
+
+        originalIndexNameDerivedSourceEnabled += restoreSuffix;
+        originalIndexNameDerivedSourceDisabled += restoreSuffix;
+        reindexFromEnabledToEnabledIndexName += restoreSuffix;
+        reindexFromEnabledToDisabledIndexName += restoreSuffix;
+        reindexFromDisabledToEnabledIndexName += restoreSuffix;
+        reindexFromDisabledToDisabledIndexName += restoreSuffix;
+
+        assertIndexBigger(originalIndexNameDerivedSourceDisabled, originalIndexNameDerivedSourceEnabled);
+        assertIndexBigger(originalIndexNameDerivedSourceDisabled, reindexFromEnabledToEnabledIndexName);
+        assertIndexBigger(originalIndexNameDerivedSourceDisabled, reindexFromDisabledToEnabledIndexName);
+        assertIndexBigger(reindexFromEnabledToDisabledIndexName, originalIndexNameDerivedSourceEnabled);
+        assertIndexBigger(reindexFromDisabledToDisabledIndexName, originalIndexNameDerivedSourceEnabled);
+        assertDocsMatch(
+            derivedSourceDisabledContext.docCount,
+            originalIndexNameDerivedSourceDisabled,
+            reindexFromEnabledToEnabledIndexName
+        );
+        assertDocsMatch(
+            derivedSourceDisabledContext.docCount,
+            originalIndexNameDerivedSourceDisabled,
+            reindexFromDisabledToEnabledIndexName
+        );
+        assertDocsMatch(
+            derivedSourceDisabledContext.docCount,
+            originalIndexNameDerivedSourceDisabled,
+            reindexFromEnabledToDisabledIndexName
+        );
+        assertDocsMatch(
+            derivedSourceDisabledContext.docCount,
+            originalIndexNameDerivedSourceDisabled,
+            reindexFromDisabledToDisabledIndexName
+        );
+    }
+
+    @SneakyThrows
+    protected void assertIndexBigger(String expectedBiggerIndex, String expectedSmallerIndex) {
+        if (isExhaustive()) {
+            logger.info("Checking index bigger assertion because running in exhaustive mode");
+            int expectedSmaller = indexSizeInBytes(expectedSmallerIndex);
+            int expectedBigger = indexSizeInBytes(expectedBiggerIndex);
+            assertTrue(
+                "Expected smaller index " + expectedSmaller + " was bigger than the expected bigger index:" + expectedBigger,
+                expectedSmaller < expectedBigger
+            );
+        } else {
+            logger.info("Skipping index bigger assertion because not running in exhaustive mode");
+        }
+    }
+
+    protected void assertDocsMatch(int docCount, String index1, String index2) {
+        for (int i = 0; i < docCount; i++) {
+            assertDocMatches(i + 1, index1, index2);
+        }
+    }
+
+    @SneakyThrows
+    protected void assertDocMatches(int docId, String index1, String index2) {
+        Map<String, Object> response1 = getKnnDoc(index1, String.valueOf(docId));
+        Map<String, Object> response2 = getKnnDoc(index2, String.valueOf(docId));
+        assertEquals("Docs do not match: " + docId, response1, response2);
+    }
+
+    @SneakyThrows
+    protected void validateVectorRecall(List<FeatureTestUtils.IndexConfigContext> indexConfigContexts) {
+        FeatureTestUtils.IndexConfigContext enabled = indexConfigContexts.get(0);
+
+        for (FeatureTestUtils.KNNVectorFieldTypeContext vectorField : enabled.collectVectorFields()) {
+            int dimension = vectorField.dimension;
+            int queryCount = Math.min(5, enabled.docCount);
+            IDVectorProducer producer = new IDVectorProducer(dimension, enabled.docCount + queryCount);
+            float[][] queryVectors = new float[queryCount][dimension];
+            for (int i = 0; i < queryCount; i++) {
+                queryVectors[i] = producer.getVector(enabled.docCount + i);
+            }
+
+            int k = Math.min(3, enabled.docCount);
+            List<List<String>> groundResults = bulkExactSearch(enabled.indexName, vectorField.fieldPath, queryVectors, k);
+            List<List<String>> testResults = bulkSearch(enabled.indexName, vectorField.fieldPath, queryVectors, k);
+            List<Set<String>> groundTruth = new ArrayList<>();
+            for (List<String> res : groundResults) {
+                groundTruth.add(new HashSet<>(res));
+            }
+            double recall = TestUtils.calculateRecallValue(testResults, groundTruth, k);
+            assertTrue("Low recall for " + vectorField.fieldPath + ":" + recall, recall >= 0.8d);
+        }
+    }
+
+    protected String getIndexName(String testPrefix, String indexPrefix, boolean addRandom) {
+        String indexName = (testPrefix + "-" + indexPrefix + getTestName()).toLowerCase(Locale.ROOT);
+        if (addRandom) {
+            indexName += randomAlphaOfLength(6);
+        }
+        return indexName.toLowerCase(Locale.ROOT);
+    }
+
+    protected Supplier<Integer> randomIntegerSupplier(long randomSeed, int min, int max) {
+        return randomIntegerSupplier(randomSeed, min, max, 1);
+    }
+
+    protected Supplier<Integer> randomIntegerSupplier(long randomSeed, int min, int max, int multipleOf) {
+        Random random = new Random(randomSeed);
+        return () -> {
+            // Calculate how many multiples fit within the range
+            int adjustedMin = (min + multipleOf - 1) / multipleOf * multipleOf;
+            int adjustedMax = max / multipleOf * multipleOf;
+
+            // Generate a random number within the adjusted range
+            int randomMultiple = random.nextInt(adjustedMin / multipleOf, (adjustedMax / multipleOf) + 1) * multipleOf;
+
+            return randomMultiple;
+        };
+    }
+
+    protected void validateDerivedSetting(String indexName, boolean expectedValue) throws IOException {
+        assertEquals(expectedValue, Boolean.parseBoolean(getIndexSettingByName(indexName, "index.knn.derived_source.enabled", true)));
+    }
 }
